backend/requirements.txt
flask==2.3.2
flask-cors==3.0.10
python-dotenv==1.0.0
pdfplumber==0.9.0
openai==0.27.8
pytesseract==0.3.10
pillow==10.0.1
pymupdf==1.22.5